+++ test/project/src/elements/input/component.config.js
@@ -1,0 +1,15 @@
+module.exports = {
+  title: 'Inputs',
+  tags: ['Atom', 'Form'],
+  context: {
+    id: 'name',
+    name: 'person[name]'
+  },
+  variants: [
+    'text.hbs',
+    'text.pug',
+    'text-required.pug',
+    'text-disabled.pug',
+    'number.pug'
+  ]
+}